separei a lat e long na base de dados e dei uns fixes nos testes

+++ miradourum/src/main/java/pt/uminho/di/aa/miradourum/services/UserService.java
@@ -17,7 +17,7 @@ public interface UserService {
      * @param id do utilizador
      * @return dados de um utilizador
      */
-    User getUserById(int id);
+    User getUserById(Long id);
 
     /**
      *
